cluster: add entry to clusterview for defining current master.
Deducing the master from an empty follow entry is not enough. In future
old master will be set to follow the new master only when the new master has
converged to the new cluster view (to avoid old master resetting itself and
failing to follow the new master). For doing this multiple members can have an
empty follow entry.