cluster: make cluster config part of the cluster view.
This is preparation work for being able to configure the cluster config.

With this patch the cluster config is managed by the leader sentinel and is
part of the clusterview, so when it's updated the clusterview version will be
increased.

This patch will (temporary) disable the ability to setup synchronous
replication since it's not possible to write the clusterconfig directly to
etcd.