proxy: stop listening only on etcd error
It makes more sense to stop listening only on communication problems with etcd
since it's the unique case when the proxy doesn't know what it should do and
shouldn't be balanced.

Instead having an empty proxyConf and closing the listening socket will exclude
the proxy from a balancer without any reason since its state is good.