Add option to provide initial cluster config
Add an `--initial-cluster-config` option to the sentinel to specify a file
containing the clusterconfig to use at cluster initialization.
This option, will be ignored on an already initialized cluster.

This is useful to define a clusterconfig without using `stolonctl config
set/patch` after the cluster is initialized.

It's also useful when there's the need to define the `InitWithMultipleKeepers`
cluster config option at first cluster initialization.