keeper: use followed node instead of master
When configuring a standby node, use the reported followed node instead
of the master. This is actually the same node but will change when
supporting cascading replication.

Also handle the case when a standby doesn't have to follow any node.
Currently these are transient cases after master election and in future
this can happen when an user will be able to define a maximum number of
standbys.