*: Move to a declarative model
This big refactor changes different concepts, logics and internal cluster data
format.

The main purpose is to move the cluster data to a fully declarative model
where every internal "object" has its own structure.
Every object has the same "structure", represented by these fields: UID,
ChangeTime, Generation, Spec and Status.
Spec is the required object state, Status contains information on the current
object status and are different for every object.

There are 3 objects: Cluster, Keeper and DB.

The user can only manipulate the desired cluster state (ClusterSpec). The
sentinel logic will update the Keeper and DB specs based on the desired cluster
state and the current state.

The keeper and DB objects are distinct also if a keeper can own only one
database because we want to be able to define when a keeper should initialize
a new db or manage the current one. We achieve this creating a new db object
with a different uid assigned to that keeper and removing the previous one.

Future evolutions permitted by this change will be:
* Implement cluster initialization using point in time recovery from a backup.
* Implement "standby" stolon clusters (where the cluster master will be
standby of an external database, like another stolon cluster).

This change breaks other things like the `stolonctl` command:
* removed the `config` command and its subcommands.
* added some new commands:
 * `init`: initialize a new cluster with a new clusterSpec.
 * `update`: replace/patch the clusterSpec.
 * `spec`: retrieve the current clusterSpec.

The sentinel api handler have been removed. This helps to add cleaner
authentication and authorization layer using only the store auth features (or
we'll have to implement both store auth and api auth causing a lot of confusion
to the user).