cluster: don't save keeper/db first error time
Saving the keeper and db first error time in the cluster data make this
subscetible to clock skews between sentinels if a new one becomes the
master.

This patch moves the first error time to a sentinel in memory value.
When a new sentinel becomes the master (or the current one is restarted)
the timers are empty so the failure detection time restarts. This isn't
a usually a big problem (just more time to declare a keeper non
healthy).