keeper: handle pg_rewind wrongly considering dbs as not forked
Two db cluster can be on the same timeline but forked at different
points in previous timelines. pg_rewind doesn't consider this case and
thinks that they are in sync and a rewind isn't needed.

Handle this case checking if, after pg_rewind, the db are not in sync and
launch a full resync.