*: Use monotonic timer
This patch adds monotonic timer support for linux. A monotonic timer is
definitely needed to avoid problems in the elapsed time calculation if
the clock jumps backward/forward due to external changes (by the user or
for any other reason).

We currently just use a clock_gettime syscall instead of the vdso since
we don't have so many calls and the perfomance gap is not noticeable and
the syscall will work on every linux arch.

The Now() method returns an int64 instead of a time.Time since it's not
a wall clock but a time since an unspecified starting point.

The timer pkg also adds a Since utility function to easily calculate
elapsed time.