cluster: implement old dead keeper removal
add a DeadKeeperRemovalInterval (default to 48 hours) to the cluster
spec and implement dead keeper removal for keepers not healthy for more
than the specified interval. Only keepers without an assigned db will be
removed:

* the keeper assigned to the current master db will never be removed
also if unhealthy or this will break the cluster leaving it without a
master.
* keeper with assigned a standby db won't be removed since they can come
back (this works well with the current logic since a db assigned to a
dead keeper will be removed and a new standby db will be assigned to a
live keeper if available)

Add a new integration test to test this behavior.