*: detect if standby can't sync due to missing wals
This patch detects if a standby won't be able to sync with master due to
missing wals.

Also if using replication slots this may happen since replication slots
are created only on the master standby so if a standby is down and a new
standby is elected as master it may not have all the wals needed by the
other standbys.

Since looks like the unique way to see if an instance cannot sync is to
parse logs for lines like:

FATAL:  could not receive data from WAL stream: ERROR:  requested WAL segment 000000010000000000000004 has already been removed

and parsing postgres logs will be error prone, we use another approach:

* Make the keepers report the older log files available in the
pg_xlog/pg_wal dir
* if a standby isn't syncing check if the required wal file name
is older than the older one available on the master. If so remove the db
from the cluster view so a new one could be readded.

Also add related integration tests.