*: add cluster spec option to disable default pg_hba.conf rules
Add a clusterspec option called `noDefaultAccessRules`. When enabled only the
minimal rules to permit the keeper to talk with the local postgres instance will
be added.

All the rules required for replication must be provided by the user in the pgHBA
option.

This is needed if users want a full control of pg_hba.conf rules (will all the
issue related to wrong/missing rules that will break the cluster).