keeper: report error when local dbuid is not correct
if the local dbuid is different than the clusterdata required one but initmode
is none something bad happened to the local keeper data. This should happen only
when the local dbstate file disappeared (ephemeral data dir, manual remove). In
such case don't just update it with the clusterdata one and continue but report an
error and stop.