Preventing panicking and improving displayed information
This commit fixes a couple of panicking bugs when empty or nil variables
are accessed without protection in some corner cases, like having an
empty `status.Cluster.MasterDBUID`, or having an empty `dbuid` passed to
the `getKeeperRoles()` function.

It also displayes more readable information in case a node has no role
in the cluster yet instead of displaying an empty string. This takes
into consideration whether information is being displayed in the default
format, or in the JSON format.